moved add client to api/v1/clients/add
