+++ fedn/network/api/interface.py
@@ -167,96 +167,6 @@ class API:
 
         return jsonify(payload)
 
-    def add_client(self, client_id, preferred_combiner, remote_addr, name, package):
-        """Add a client to the network.
-
-        :param client_id: The client id to add.
-        :type client_id: str
-        :param preferred_combiner: The preferred combiner for the client.If None, the combiner will be chosen based on availability.
-        :type preferred_combiner: str
-        :return: A json response with combiner assignment config.
-        :rtype: :class:`flask.Response`
-        """
-        if package == "remote":
-            package_object = self.statestore.get_compute_package()
-            if package_object is None:
-                return (
-                    jsonify(
-                        {
-                            "success": False,
-                            "status": "retry",
-                            "message": "No compute package found. Set package in controller.",
-                        }
-                    ),
-                    203,
-                )
-            helper_type = package_object["helper"]
-        else:
-            # Else package is "local":
-            helper_type = ""
-
-        # Assign client to combiner
-        if preferred_combiner:
-            combiner = self.control.network.get_combiner(preferred_combiner)
-            if combiner is None:
-                return (
-                    jsonify(
-                        {
-                            "success": False,
-                            "message": f"Combiner {preferred_combiner} not found or unavailable.",
-                        }
-                    ),
-                    400,
-                )
-        else:
-            combiner = self.control.network.find_available_combiner()
-            if combiner is None:
-                return (
-                    jsonify({"success": False, "message": "No combiner available."}),
-                    400,
-                )
-
-        client_config = {
-            "client_id": client_id,
-            "name": name,
-            "combiner_preferred": preferred_combiner,
-            "combiner": combiner.name,
-            "ip": remote_addr,
-            "status": "available",
-            "package": package,
-        }
-        # Add client to network
-        self.control.network.add_client(client_config)
-
-        payload = {
-            "status": "assigned",
-            "host": combiner.address,
-            "fqdn": combiner.fqdn,
-            "package": package,
-            "ip": combiner.ip,
-            "port": combiner.port,
-            "helper_type": helper_type,
-        }
-        return jsonify(payload)
-
-    def get_model(self, model_id: str):
-        result = self.statestore.get_model(model_id)
-
-        if result is None:
-            return (
-                jsonify({"success": False, "message": "No model found."}),
-                404,
-            )
-
-        payload = {
-            "committed_at": result["committed_at"],
-            "parent_model": result["parent_model"],
-            "model": result["model"],
-            "session_id": result["session_id"],
-        }
-
-        return jsonify(payload)
-
     def get_client_config(self, checksum=True):
         """Get the client config.
